refactor: move authorlist to its own component, test author component
